@@ -15,17 +15,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.10",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "^6.0.3",
-    "@angular/compiler": "^6.0.3",
-    "@angular/core": "^6.0.3",
-    "@angular/forms": "^6.0.3",
-    "@angular/http": "^6.0.3",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "^6.0.3",
-    "@angular/platform-browser-dynamic": "^6.0.3",
-    "@angular/router": "^6.0.3",
+    "@angular/animations": "~6.1.10",
+    "@angular/cdk": "~7.3.7",
+    "@angular/common": "~6.0.3",
+    "@angular/core": "~6.0.3",
+    "@angular/flex-layout": "~7.0.0-beta.24",
+    "@angular/forms": "~6.0.3",
+    "@angular/http": "~6.0.3",
+    "@angular/material": "^6.4.7",
+    "@angular/platform-browser": "~6.0.3",
+    "@angular/platform-browser-dynamic": "~6.0.3",
+    "@angular/router": "~6.0.3",
     "@mdi/font": "^3.6.95",
     "core-js": "^2.5.4",
     "rxjs": "^6.0.0",
@@ -33,8 +33,9 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.6.8",
-    "@angular/cli": "^6.0.8",
-    "@angular/compiler-cli": "^6.0.3",
+    "@angular/cli": "~7.3.8",
+    "@angular/compiler": "~7.2.14",
+    "@angular/compiler-cli": "~7.2.14",
     "@angular/language-service": "^6.0.3",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
@@ -50,6 +51,6 @@
     "protractor": "~5.3.0",
     "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~2.7.2"
+    "typescript": "~3.1"
   }
 }
