@@ -1,6 +1,9 @@
 import {Component, OnInit, HostListener, ViewChild, ElementRef, Output, EventEmitter} from '@angular/core';
 import {FormControl, FormGroup, Validators} from '@angular/forms';
 import { debounceTime } from 'rxjs/operators';
+import {PresetsService} from '../services/presets.service';
+import {Observable} from 'rxjs/index';
+import {FilterService} from '../services/filter.service';
 
 export enum SimpleState {
   Any = 'Any',
@@ -14,17 +17,14 @@ export enum SimpleState {
   styleUrls: ['./side-menu.component.scss']
 })
 export class SideMenuComponent implements OnInit {
-  @ViewChild('searchInput') searchInput: ElementRef;
 
-  defaultPresetName = 'New preset';
-  searchExample = 'progress, background, experience';
-  displayHelp = false;
+  presetName: string;
+  presets: Observable<any[]>;
 
   alphaFilterItems = [SimpleState.Any, SimpleState.Yes, SimpleState.No];
   sliceFilterItems = [SimpleState.Any, SimpleState.Yes, SimpleState.No];
 
   form = new FormGroup({
-      search: new FormControl(),
       xMin: new FormControl('', [Validators.min(0)]),
       xMax: new FormControl('', [Validators.min(0)]),
       yMin: new FormControl('', [Validators.min(0)]),
@@ -35,7 +35,6 @@ export class SideMenuComponent implements OnInit {
   );
 
   defaultPreset = {
-    search: '',
     alpha: SimpleState.Any,
     slice: SimpleState.Any,
     xMin: '',
@@ -44,15 +43,9 @@ export class SideMenuComponent implements OnInit {
     yMax: '',
   };
 
-  presetName: string;
-  presets = [];
-
-  @Output() filterChange = new EventEmitter();
-
-  private _skipHelpHiding = false;
-
-  constructor() {
+  constructor(private _presetService: PresetsService, private _filterService: FilterService) {
     this.resetForm();
+    this.presets = _presetService.presetsSource;
   }
 
   ngOnInit() {
@@ -61,42 +54,22 @@ export class SideMenuComponent implements OnInit {
         debounceTime(100)
       )
       .subscribe(ch => this.onChange());
-    this.loadPresetFromStorage();
+
+    this._presetService.currentPreset.subscribe(preset => this.setPreset(preset));
   }
 
   @HostListener('window:keydown', ['$event'])
   keydown(event: KeyboardEvent) {
-    // F3 and ctrl+f
-    if (event.keyCode === 114 || (event.ctrlKey && event.keyCode === 70)) {
-      event.preventDefault();
-      this.selectSearchField();
-    }
     // ctrl+S
-    else if(event.ctrlKey && event.keyCode === 83){
+    if (event.ctrlKey && event.keyCode === 83) {
       this.savePreset();
       event.preventDefault();
     }
   }
 
-  @HostListener('window:click', ['$event'])
-  click(event) {
-    if (this._skipHelpHiding) {
-      this._skipHelpHiding = false;
-      return;
-    }
-    if (this.displayHelp) {
-      this.displayHelp = false;
-    }
-  }
-
-
-  selectSearchField() {
-    this.searchInput.nativeElement.select();
-  }
-
   onChange() {
     console.log('changed');
-    this.filterChange.emit(this.getCurrentPreset());
+    this._filterService.filterSprites(this.form.value);
   }
 
   changeFilter(field: string, value: any) {
@@ -104,43 +77,11 @@ export class SideMenuComponent implements OnInit {
     this.onChange();
   }
 
-  showHelp() {
-    this.displayHelp = !this.displayHelp;
-    this._skipHelpHiding = true;
-  }
-
-  setExample() {
-    this.form.controls.search.setValue(this.searchExample);
-  }
-
   savePreset() {
-    const preset = this.getCurrentPreset();
-    let newPresetName = this.presetName || this.form.controls.search.value || this.defaultPresetName;
-    newPresetName = newPresetName.trim();
-    if(newPresetName.length > 20) {
-      newPresetName = newPresetName.slice(0, 20) + '...';
-    }
-    let count = 1;
-    let temp = newPresetName;
-    while (this.presets.find(pr => pr.presetName === temp)) {
-      count++;
-      temp = `${newPresetName} (${count})`;
-    }
-    preset['presetName'] = temp;
+    const preset = this.form.value;
+    this._presetService.updateCurrentState(preset);
+    this._presetService.savePreset(this.presetName);
     this.presetName = undefined;
-    this.presets.push(preset);
-
-    this.savePresetsInStorage();
-  }
-
-  getCurrentPreset() {
-    const preset = {};
-    for (const prop in this.defaultPreset) {
-      if (this.defaultPreset.hasOwnProperty(prop)) {
-        preset[prop] = this.form.controls[prop].value;
-      }
-    }
-    return preset;
   }
 
   setPreset(preset) {
@@ -151,24 +92,15 @@ export class SideMenuComponent implements OnInit {
     }
   }
 
-  removePreset(index) {
-    this.presets.splice(index, 1);
-    this.savePresetsInStorage();
+  loadPreset(index) {
+    this._presetService.loadPreset(index);
   }
 
-  loadPreset(index) {
-    this.setPreset(this.presets[index]);
+  removePreset(index) {
+    this._presetService.removePreset(index);
   }
 
   resetForm() {
     this.setPreset(this.defaultPreset);
-  }
-
-  savePresetsInStorage() {
-    localStorage.setItem('presets', JSON.stringify(this.presets));
-  }
-
-  loadPresetFromStorage() {
-    this.presets = JSON.parse(localStorage.getItem('presets')) || [];
   }
 }
